+++ src/components/News/index.ts
@@ -1,0 +1,3 @@
+import News from './News';
+
+export { News };
