+++ src/components/Profile/MyPosts/addPostForm/index.ts
@@ -1,0 +1,3 @@
+import AddPostForm from './AddPostForm';
+
+export { AddPostForm };
