+++ src/components/Profile/MyPosts/index.ts
@@ -1,0 +1,4 @@
+import MyPosts from './MyPosts';
+import MyPostsContainer from './MyPostsContainer';
+
+export { MyPosts, MyPostsContainer };
