+++ src/components/Profile/ProfileInfo/ProfileInfo.tsx
@@ -1,78 +1,78 @@
 import { FC, useState } from 'react';
 import { ProfileType } from '../../../@types/types.ts';
 import wallpaperImg from '../../../images/profile_wallpaper.jpg';
-import { Preloader } from '../../common/Preloader/Preloader';
-import ProfileDescription from './ProfileData/ProfileDescription.tsx';
-import ProfileDescriptionForm from './ProfileData/ProfileDescriptionForm.tsx';
-import ProfilePhoto from './ProfileData/ProfilePhoto.tsx';
+import { Preloader } from '../../common/preloader/Preloader.tsx';
+import { ProfileDescription } from './profileData/profileDescription';
+import { ProfileDescriptionForm } from './profileData/profileDescriptionForm';
+import { ProfilePhoto } from './profileData/profilePhoto';
 import st from './ProfileInfo.module.css';
-import ProfileStatusWithHooks from './ProfileStatusWithHooks';
+import { ProfileStatusWithHooks } from './profileStatus';
 
 type Props = {
-  profile: ProfileType | null;
-  status: string;
-  updateUserStatus: (status: string) => void;
-  isOwner: boolean;
-  updatePhoto: (file: File) => void;
-  updateProfileData: (profileData: ProfileType) => void;
-  errorMessage: string;
+    profile: ProfileType | null;
+    status: string;
+    updateUserStatus: (status: string) => void;
+    isOwner: boolean;
+    updatePhoto: (file: File) => void;
+    updateProfileData: (profileData: ProfileType) => void;
+    errorMessage: string;
 };
 
 const ProfileInfo: FC<Props> = ({
-  profile,
-  status,
-  updateUserStatus,
-  isOwner,
-  updatePhoto,
-  updateProfileData,
-  errorMessage,
+    profile,
+    status,
+    updateUserStatus,
+    isOwner,
+    updatePhoto,
+    updateProfileData,
+    errorMessage,
 }) => {
-  const [editMode, setEditMode] = useState(false);
+    const [editMode, setEditMode] = useState(false);
 
-  if (!profile) {
-    return <Preloader />;
-  }
+    if (!profile) {
+        return <Preloader />;
+    }
 
-  const onSubmit = (profileData: ProfileType) => {
-    setEditMode(false);
-    updateProfileData(profileData);
-  };
+    const onSubmit = (profileData: ProfileType) => {
+        setEditMode(false);
+        updateProfileData(profileData);
+    };
 
-  return (
-    <div className={st.profileInfo}>
-      <div className={st.image}>
-        <img
-          src={wallpaperImg}
-          alt=""
-        />
-      </div>
+    return (
+        <div className={st.profileInfo}>
+            <div className={st.image}>
+                <img
+                    src={wallpaperImg}
+                    alt=""
+                />
+            </div>
 
-      <div className={st.profile}>
-        <ProfilePhoto
-          profile={profile}
-          isOwner={isOwner}
-          updatePhoto={updatePhoto}
-        />
-        {errorMessage && <div className={st.errorMessage}>{errorMessage}</div>}
-        {editMode ? (
-          <ProfileDescriptionForm
-            profile={profile}
-            onSubmit={onSubmit}
-          />
-        ) : (
-          <ProfileDescription
-            profile={profile}
-            isOwner={isOwner}
-            setEditMode={setEditMode}
-          />
-        )}
-      </div>
-      <ProfileStatusWithHooks
-        status={status}
-        updateUserStatus={updateUserStatus}
-      />
-    </div>
-  );
+            <div className={st.profile}>
+                <ProfilePhoto
+                    profile={profile}
+                    isOwner={isOwner}
+                    updatePhoto={updatePhoto}
+                />
+                {errorMessage && <div className={st.errorMessage}>{errorMessage}</div>}
+                {editMode ? (
+                    <ProfileDescriptionForm
+                        profile={profile}
+                        onSubmit={onSubmit}
+                    />
+                ) : (
+                    <ProfileDescription
+                        profile={profile}
+                        isOwner={isOwner}
+                        setEditMode={setEditMode}
+                    />
+                )}
+            </div>
+            <ProfileStatusWithHooks
+                status={status}
+                updateUserStatus={updateUserStatus}
+            />
+        </div>
+    );
 };
 
 export default ProfileInfo;
