+++ src/components/Profile/ProfileInfo/index.ts
@@ -1,0 +1,3 @@
+import ProfileInfo from './ProfileInfo';
+
+export { ProfileInfo };
