+++ src/components/Profile/ProfileInfo/profileData/contact/index.ts
@@ -1,0 +1,3 @@
+import Contact from './Contact';
+
+export { Contact };
