+++ src/components/Profile/ProfileInfo/profileData/contactForm/ContactForm.tsx
@@ -1,0 +1,14 @@
+import { createField } from 'components/common/formsControls/createField.tsx';
+import { Input } from '../../../../common/formsControls/FormControls.tsx';
+import st from './ContactForm.module.css';
+
+const ContactForm = ({ contactName }: { contactName: string }) => {
+    return (
+        <div className={st.contact}>
+            <b>{contactName}:</b>
+            {createField(`contacts.${contactName}`, '', Input, contactName, undefined)}
+        </div>
+    );
+};
+
+export default ContactForm;
