+++ src/components/Profile/ProfileInfo/profileData/contactForm/index.ts
@@ -1,0 +1,3 @@
+import ContactForm from './ContactForm';
+
+export { ContactForm };
