+++ src/components/Profile/ProfileInfo/profileData/profileDescription/ProfileDescription.module.css
@@ -1,0 +1,61 @@
+.image {
+  position: relative;
+  overflow: hidden;
+  width: 100%;
+  height: 300px;
+}
+
+.image img {
+  position: absolute;
+  top: 70%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 100%;
+  height: 450px;
+  object-fit: cover;
+}
+
+.profile {
+  padding: 15px;
+}
+
+.ava {
+  margin: 10px;
+}
+
+.ava img {
+  height: 220px;
+  width: 220px;
+  object-fit: cover;
+}
+
+.description {
+  font-size: 20px;
+}
+
+.description .info h2 {
+  color: black;
+}
+.errorMessage {
+  border: 1px solid red;
+  color: red;
+  margin-bottom: 15px;
+  padding: 3px;
+}
+
+.description .info .lookingForAJob {
+  display: flex;
+  align-items: center;
+}
+
+.contact {
+  padding-left: 15px;
+}
+
+.profileStatus {
+  margin-left: 20px;
+}
+
+.profileStatus input {
+  font-size: 15px;
+}
