+++ src/components/Profile/ProfileInfo/profileData/profileDescription/index.ts
@@ -1,0 +1,3 @@
+import ProfileDescription from './ProfileDescription';
+
+export { ProfileDescription };
