+++ src/components/Profile/ProfileInfo/profileData/profileDescriptionForm/index.ts
@@ -1,0 +1,3 @@
+import ProfileDescriptionForm from './ProfileDescriptionForm';
+
+export { ProfileDescriptionForm };
