+++ src/components/Profile/ProfileInfo/profileData/profilePhoto/index.ts
@@ -1,0 +1,3 @@
+import ProfilePhoto from './ProfilePhoto';
+
+export { ProfilePhoto };
