+++ src/components/Profile/ProfileInfo/profileStatus/index.ts
@@ -1,0 +1,4 @@
+import ProfileStatus from './ProfileStatus';
+import ProfileStatusWithHooks from './ProfileStatusWithHooks';
+
+export { ProfileStatus, ProfileStatusWithHooks };
