+++ src/components/Profile/index.ts
@@ -1,0 +1,4 @@
+import Profile from './Profile';
+import ProfileContainer from './ProfileContainer';
+
+export { Profile, ProfileContainer };
