+++ src/components/Users/User/User.module.css
@@ -1,0 +1,82 @@
+.users {
+  max-width: 900px;
+  margin: 50px auto;
+  font-size: 20px;
+}
+
+.user {
+  display: grid;
+  margin: 15px;
+  justify-items: center;
+  align-items: center;
+  grid-template-columns: 1fr 5fr;
+}
+
+.ava {
+  width: 80px;
+  height: 80px;
+  margin: 5px;
+}
+
+.ava img {
+  width: 100%;
+  border-radius: 50%;
+  background-color: rgb(126, 196, 125);
+}
+
+.button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.button > button {
+  width: 80px;
+  background-color: bisque;
+  height: 2em;
+  border: 0;
+  outline: 0;
+  border-radius: 5px;
+  min-width: 10%;
+  font-size: 15px;
+  cursor: pointer;
+}
+
+.button > button:hover {
+  background-color: rgb(246, 217, 180);
+  box-shadow: 0px 1px 6px #232323;
+  transition: 0.2s;
+  object-fit: contain;
+}
+
+.button > button:active {
+  transform: scale(0.95);
+}
+
+.biography {
+  display: grid;
+  grid-template-columns: 3fr 1fr;
+  width: 100%;
+  height: 100%;
+  border: solid 2px #fff;
+  border-radius: 10px;
+}
+
+.about {
+  display: grid;
+  padding: 5px;
+  align-items: center;
+  grid-template-rows: repeat(2, 1fr);
+}
+
+.name {
+  font-size: 25px;
+  font-weight: bold;
+}
+
+.location {
+  display: grid;
+  align-items: center;
+  justify-items: center;
+  padding: 5px;
+}
