+++ src/components/Users/User/index.ts
@@ -1,0 +1,3 @@
+import User from './User';
+
+export { User };
