+++ src/components/Users/index.ts
@@ -1,0 +1,4 @@
+import Users from './Users';
+import UsersContainer from './UsersContainer';
+
+export { Users, UsersContainer };
